USB audio: Fix building on x86_64
* fix x86_64 build errors and warnings pointed out by Diver;
* fix warning building with gcc2.

+++ src/add-ons/kernel/drivers/audio/usb/Driver.cpp
@@ -254,7 +254,7 @@ publish_devices()
 	MutexLocker driverLock;
 
 	int32 deviceCount = 0;
-	for (int32 i = 0; i < MAX_DEVICES; i++) {
+	for (size_t i = 0; i < MAX_DEVICES; i++) {
 		if (gDevices[i] == NULL)
 			continue;
